[*] Support command functions that don't have a return value.
The interface used to to communicate what should be run from a command
store had a return value specified. This required the users were forced
to return something even when it did not make sense.

Often times the forced return type was mitigated by just returning
`null`. This made the use of the api annoying in places.

We now support a method where the user can specify a command function
with a different interface. Here no return value is required. Under the
hood this is handled by turning a 'consumer' like method into a
'function' like method by adding a `return null`.